move `exposeNativeInterface`
it was in the wrong chunk
